Remove proxy support, and bump version
The proxy may be causing problems, so we're avoiding it now. Put the panos you
want to load behind a web server that will use allow-origin headers so the
javascript can load them. If that doesn't work, put the panos inside this
activity itself. Unzip the activity's zip file, put the panos in the
webapp/textures directory, and zip it up again before loading into ispaces.
Alternatively, once the live activity is deployed, find its UUID in the ispaces
interface, and copy the panos to
${your_controller's_home}/controller/activities/installed/${activity's_UUID}/install/webapp/textures
directory. In either of these latter two instances, the panos' URLs will be '/textures/${pano_name}'